Version 0.97
In order to make the game more challenging, more complicated algorithm is used to generate maze, and determine target. However, this slow the script down drastically, especially when generating large maze. If you find it too slow, you may want to use older version(0.96) of the script.
Now the constrain of row is [5,30] and column is [5,40]. Please choose appropriate size according to the size of your vim window and the code's running time you can tolerate.